+++ gladius/consts.py
@@ -96,7 +96,5 @@ EVENT_HANDLER_EVENT_TYPE_MAP: dict[str, str] = {
     'onvisibilitychange': 'visibilitychange',
 
     # custom
-    '_ontextchange': '_textchange',
-    '_ontablechange': '_tablechange',
     '_oncontentchange': '_contentchange',
 }